refactor(gateway): extract LightClientReadOnly (#106 phase 4)
Pulls the six read-only JNI passthroughs out of GatewayRepository:
- getPeers / getTipHeader / getScripts / callRpc (NodeStatus diagnostic)
- currentTipNumberOrZero (parsed tip block)
- getCurrentEpoch (DAO epoch math)

GatewayRepository drops from 2087 to ~2070 lines. The repository
keeps thin delegation shims so NodeStatusViewModel and the DAO
flow paths are unchanged. None of the extracted methods mutate
state, which made this the least-risky remaining cluster — the
issue's own "out of scope" cautioning applies to the bigger
send/network-switch pipeline that's still on GatewayRepository.

Cumulative #106 progress across four phases:
- Phase 1 #247 SyncCoordinator (~370 lines)
- Phase 2 #248 DaoHeaderResolver (~150 lines)
- Phase 3 #250 DaoDepositReader (~290 lines)
- Phase 4 (this) LightClientReadOnly (~85 lines)
- GatewayRepository.kt: 2629 → ~2070 (~21%)

611/611 tests pass.

Refs #106
